refactor: check for finite float/double

+++ src/main/java/me/chayapak1/chomens_bot/util/TimestampUtilities.java
@@ -20,7 +20,7 @@ public class TimestampUtilities {
 
         if (hourString != null) time += (long) Integer.parseInt(hourString) * 60 * 60 * 1000;
         time += (long) Integer.parseInt(minuteString) * 60 * 1000;
-        time += Double.parseDouble(secondString) * 1000.0;
+        time += (long) (Double.parseDouble(secondString) * 1000.0);
 
         return time;
     }
